+++ src/js/index.js
@@ -1,15 +1,10 @@
 import { popularMovies } from './popular-movies-home';
-import {
-  getDataAndCreatePagination,
-  createPagination,
-  goToPage,
-} from './pagination';
 
 const gallery = document.querySelector('.gallery-cards');
 let page = 1;
+global.popularMovies = popularMovies;
 
 document.addEventListener('DOMContentLoaded', () => {
   popularMovies(1);
-
-  getDataAndCreatePagination();
+  // getDataAndCreatePagination();
 });
